requirements.txt
pyrogram==1.4.16
TgCrypto==1.2.5
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
